update sample project
